# main/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple Twitter-like REST API.",
        "title": "Twitter-like API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1/",
    "paths": {
...
        "/protected/ws": {
            "get": {
                "description": "Upgrade HTTP connection to WebSocket for real-time notifications. Requires JWT token in query param or Authorization header.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "websocket"
                ],
                "summary": "Connect to WebSocket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Access Token (optional if header present)",
                        "name": "token",
                        "in": "query"
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized or invalid token",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Use \"Bearer {access_token}\" format.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}